+++ app/src/main/kotlin/io/github/maxlyth/hapaneld/device/profile/DataDeviceProfile.kt
@@ -33,6 +33,7 @@ class DataDeviceProfile internal constructor(
     }
     override val appCanSu = document.platform.appCanSu
     override val hasRecents = document.platform.hasRecents
+    override val hasNativeNavbar = document.platform.hasNativeNavbar
     override val ledMechanism = when (document.hardware.led.mechanism) {
         "rk3576-ioctl" -> LedMechanism.RK3576_IOCTL
         "rk3576-ioctl-daemon" -> LedMechanism.RK3576_IOCTL_DAEMON